fix: overlays not visible when nesting Vuetify and VuetifyTemplate
Extract vue render functions out of widget views so they can be
used without first having to make a view. The created view for a
nested Vuetify or VuetifyTemplate widget was preventing the
overlay (e.g. the list of a Select) from being displayed.

fixes #18
